requirements.txt
# Web Framework
flask==3.0.0
werkzeug==3.0.1

# MongoDB Driver
pymongo==4.6.1

# YAML Configuration
pyyaml==6.0.1

# Elasticsearch Client
elasticsearch>=8.0.0,<9.0.0

# Additional dependencies (automatically installed but listed for clarity)
click>=8.1.7
itsdangerous>=2.1.2
jinja2>=3.1.2
markupsafe>=2.1.3
blinker>=1.7.0
